# chore(performance): Improve performance for adding RemoveExtraSlash flag